requirements.txt
Django==4.2.3
Pillow==8.4.0
tweepy==3.10.0
